+++ src/markdown/ChangeLog.md
@@ -1,23 +1,31 @@
 ## Changelog
 
-### v1.03 (22 November 2021)
+### v1.0.4 (28 November 2021)
+
+* Added Relics into the tool, based on data sourced from the Siralim Ultimate Compendium.
+* Added a reset button.
+* Added a randomise button.
+
+Thank you to EmptyPalms for the suggestions!
+
+### v1.0.3 (22 November 2021)
 
 * Added data from Siralim Ultimate Compendium v0.12.12.
 * Note that monster sprites and stats for the new creatures are not currently present, and will be added once Siralim Ultimate API is updated.
 * Unfortunately builds from the previous patch are broken - sorry about that. It shouldn't happen in future though unless the trait names, creature names, or family names change again.
 
-### v1.02 (1 October 2021)
+### v1.0.2 (1 October 2021)
 
 * Major design overhaul.
 * Added the ability to select a specialization and anointments.
 * Added monster stats to the monster selection table.
 * The monster selection table can now be sorted by clicking on the headings.
 
-### v1.01 (25 September 2021)
+### v1.0.1 (25 September 2021)
 
 * Added creature sprites, sourced from the Siralim Ultimate API.
 * Tool now allows you to upload your build directly from the export from the game.
 
-### v1.0 (24 September 2021)
+### v1.0.0 (24 September 2021)
 
 * Released.